Member Card components moved to another folder
